requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sock==0.7.0
google-genai
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
flask-session==0.5.0
